+++ src/core/hash.js
@@ -5,6 +5,9 @@
  * @returns {number} Hash value.
  */
 function hashCode(str) {
+    if (str === null || str === undefined) {
+        return 0;
+    }
     let hash = 0;
     for (let i = 0, len = str.length; i < len; i++) {
         hash = ((hash << 5) - hash) + str.charCodeAt(i);
