+++ src/scene/shader-lib/chunks/chunks.js
@@ -56,8 +56,10 @@ import gamma2_2PS from './common/frag/gamma2_2.js';
 import gles3PS from '../../../platform/graphics/shader-chunks/frag/gles3.js';
 import gles3VS from '../../../platform/graphics/shader-chunks/vert/gles3.js';
 import glossPS from './standard/frag/gloss.js';
+import gsplatCoreVS from './gsplat/vert/gsplatCore.js';
 import gsplatEvalPS from './gsplat/frag/gsplatEval.js';
 import gsplatMainPS from './gsplat/frag/gsplatMain.js';
+import gsplatMainVS from './gsplat/vert/gsplatMain.js';
 import iridescenceDiffractionPS from './lit/frag/iridescenceDiffraction.js';
 import iridescencePS from './standard/frag/iridescence.js';
 import iridescenceThicknessPS from './standard/frag/iridescenceThickness.js';
@@ -258,8 +260,10 @@ const shaderChunks = {
     gles3PS,
     gles3VS,
     glossPS,
+    gsplatCoreVS,
     gsplatEvalPS,
     gsplatMainPS,
+    gsplatMainVS,
     iridescenceDiffractionPS,
     iridescencePS,
     iridescenceThicknessPS,
